Added Cinemachine
Added Cinemachine changed the settings and added Confiner so that the camera can't go past a certain area.
Also got rid of Camera Controller in the main camera since we don't need it anymore.
Got rid of some extra comments and print from the player movement that we don't need anymore.